Baby steps.
lua checks can now be written.  Simple wrappers for check manipulation are
in place.  I need to figure out how to make lua resume into a coroutine and
invoke lua_error on resumption (so we can force a timeout).

noit.sleep(number) was implemanted as a proof-of-concept coro... It works!

Need to do sockets next -- will be slightly more complicated, but I do not
see it being traumatic.

refs #28