This is a hack, refs #87.
I don't like the way this is done.  The problem here is that the log
log string is fixated for the first log stream and output vectors
use their own.  So, you get all or nothing and can't debug at lower
tiers.

Ideally, the log API must change to accept (time, file, line) for
all the writes.  That way they can do what they like with it and
we can create a maintainer/debug logger that brings crazy detailed
information.

We only have a few loggers now, it should be very easy.