correctly support namespaces in disjoint (shatter) documents.
The shatter code glues multiple documents together and as such
the namespace we need may not be in the super-parent even after
we add it to our specific node.  So, after we miss out of the
super-parent, we should check to see if we've added it to the
node we're about to add it to.