Added Execution Path To External, Plus Bug Fixes
Added a new module config variable, "path", to the External check.
This will specify a directory to run commands out of; if an individual
check tries to specify a command outside of this path, it will not
be executed.

In addition, fixed a bug where the "group" permissions were not being
set properly and changed the check statuses of failed checks to be
more descriptive/useful.