Added Nagios Parsing Ability To External Check
Added the capability to parse Nagios checks meaningfully to the
external check. There is now a module config, "nagios_regex", that
contains the regex to parse out the output from Nagios scripts. If
"NAGIOS" is entered as the output_extract parameter, noitd will enter
a special Nagios parsing mode that will pull out the state and any
associated metrics from the Nagios script.