Fix Broken DNS TTL Handling
TTL handling for DNS lookups was somewhat broken. We would do two DNS
lookups for each target - an IPv4 and an IPv6. We would end up setting
the TTL to whatever the second result to come back was; so, if we got a
valid IPv4 result, but an invalid IPv6 result, we would override the
TTL with the default value, 60 seconds.

Now, only set the default value if the other resolver is still in-flight...
This will fix the issue where the second result comes back and is invalid,
causing us to ignore the first TTL result.