Make runtests work on OmniOS and avoid compilation if poassible.
This is a truly digusting amount of work to get npm to "just work"
if you're on a system that doesn't have all your required shit dumped
in /usr/lib and /usr/include.

1) In trying to get protobuf to compile I introduced an evil sed hack
from a parent Makefile to assist in extracting includes and linker paths.

2) In trying to avoid compiling protobuf every time and discovering the
ruined atrocity that is 'npm link' (WTF do you mean I need permissions
to rebuild the perfectly functional, globally-install copy?), I added a
version-blind hack to direct link the global install.  I also marked this
package as private so that npm install would STFU.

At this point, if there are globally installed protobuf and fq modules
one can dispense with npm compiling anything as a prerequisite of
running the test suite.