Fix dns callbacks that attempt to resume the caller.
We could attempt a resume of ourselves in a C func when we call
dns_timeouts within the lookup C func.  We need to note that we're
inside a lua C call and note the stack state for return instead of
resuming and then return it directly otherwise yield as we did.