always name mtev_console structure types.
This avoids a C++ warning, X has a field X::Y whose type uses the
anonymous namespace. In this case, we avoid the error by requiring
all structure types to have a name before the open-brace, but
signal to clients that these structure types should not be used
directly by prefixing the name with `_`.