Fix cross-thread trigger under kqueue.
On kqueue we only request changes to event flags. If we were calling
cross thread the kqueue on that thread had no registrations, so we
need to zero them out in the source thread and then force them to
be put in place post-callback in the target thread.  The requires
knowing that we've been called in such a fashion and we've introduced
the EVENTER_CROSS_THREAD_TRIGGER mask to indicate such a calling
condition.