Avoid SSL_SESSION_free (#39)
SSL_get1_session is supposed to increment the refcnt.  Other things can
negatively effect the SSL_SESSION itself making it dangerous to free even
the refcnt'd copy (crashes on OS X).  There is only one consumer of this
and we can avoid this by simply requiring that no one use it "later."

Expose a :release() method on the session in lua that null our reference
and make the GC null its reference.  Then change the return convention
to not increment the refcnt.