Better Error Handling Before Aborts
When we abort, rather than just aborting, log what happens via the new
MTEV_LOG_AND_ABORT() function. This will set the log to synchronous, log
the message, then abort. This will avoid the issue where we queue
something to be written to a log but exit before it flushes.