Add FQ-modules-only install target
Facilitates packaging, where we want to deliver the optional FQ
modules separately. Only one change is present for default builds:
mtev_fq.h is no longer installed unless we are building with
FQ support. This header is useless without fq.so anyway.

As part of this fix we disambiguate the Lua-related modules to reduce
confusion. There is one module that provides Lua functionality to
libmtev apps (LUA4MTEV) and one that provides libmtev functionality to
Lua apps (MTEV4LUA).

Also fixed is a bug with cleaning where we were not cleaning up in the
src/modules directory.