Add Locking To Certain Hash Tables
Did an audit to initialize certain hash tables as locking. I didn't mess
with any tables that manually have locks around them, and I tried to be
conservative with initializing the locks so that we don't lock
unnecessarily - this may require tweaking at a later point, but this is
a good start.