Accommodate hwloc's destructive processor_bind() on Solaris/Illumos.
hwloc should spin up a thread and bind the new thread, then join; it
does not.  So, we use the ctor to induce all this purely evil stuff
to happen early enough in the boot process to not screw over every
other thread or process binding that might be serviced.