Support multiple concurrency watchdog timers.
Since making the event loop multi-threaded, it is possible for
a thread other than the master thread to hang without triggering
a watchdog timeout.  This provides an API where any consumer
can create a new watchdog and maintain it such that the parent
will notice inactivity.  This commit does not enable per-thread
watchdogs in the eventer as that change would conflict with some
other in-flight changes; it will come later.