Fix mtev.tojson
Switch to use more idiomatic lua stack references.
When iterating a table that has both integer and string keys
we must not ask for conversion to the iteration will break.
Specifically, we must not ask for the integer ID as a string,
we must get it as an integer and make our own string.