Make Linux epoll impl more resilient to confused add/del/mod.
This change isn't ideal as I think it accommodates other bugs, but
it would appear that we can remove and fd and then re-trigger it
with a mask and an event in the master_fds.  This causes us to
attempt to add when we should modify (or mod when we should add).
This simply provides an explicit fallback if the approach we take
errors b/c it was backwards, we try the other operation.