More coverity fixes (#303)
* More coverity fixes
Fixed a few more issues in coverity. Need to push and merge this so I can do a git pull and Fix a few more remaining changes.
Reviews appreciated.

* Fix
Fixed not adding to the already present if statement.

* Spacing fix
Removed a trailing space and added spaces around the OR in an if statement.