Fix Potential Race Condition In Log Setup
If we try to find a log and fail, we attempt to create it; however,
another thread MAY have already created the log, which would lead to
storing the "new" log failing, which would lead to a crash. If we fail
to open a log, rather than crashing, we should try to retrieve it again;
if THAT fails, we should just assert out rather than dying by trying to
access a null pointer.