Fix Errors In mtev_reverse_socket_handler
If there is outgoing data, we need to write it out in a loop rather than
just writing one; otherwise, we may not set the mask properly and can
get hung.

Also - don't always write all data - try to read/write 500 entries at a
time. If you hit that limit, return a full mask so we roll back and try
again.