Fix race in lua subsystem w.r.t. coroutine GC.
It is possible that a coroutine is "removed" from the mtev
registry and and yet garbarge collection is delayed enough
that it still executed in the lua VM.  This would cause the
revivification of the resume_info in the mtev coro state
tracker and desperately confuse things.  Events that are
solely owned (and are active) within a coro that is canceled
can fire their callbacks as well.  In these changes we
see a few things:

 * resume info states know how to create same time infos.
 * a find function for resume info was added that does not
   autovivify resume info when missing.
 * all callbacks from the event system that land in a lua
   state that has no resume info are cleaned up; lua error.
 * better safety is added for events removed on-demand that
   then later get removed via resume info cleanup.