Make lua GC on next eventer tick. (#460)
* Make lua GC on next eventer tick.

We can be in a lua call triggering into C completing and decide "we're done"
with the coroutine that is servicing the request.  From lua's perspective
we are done, but it is littered through the calling frames and is very likely
going to be used before we unwind.  So, instead of GC in place, we make
mtev_lua_gc schedule a GC on next tick which makes for a clean (empty) call
stack and greatly simplifies the next of lua/C call stacks.

This also places the threads in the global registry as recommended instead
of an "mtev_coros" global table.

* coro harness