Fix ACO rest calls.
ACO event resumption reall needs to start at an event callback level.
This new approach floats the event, schedules a continue at next tick.
The continuation starts the coroutine which captures the original
callback and closure, aco-izes the event and call the original bits.
At this point, the aco should be completely eventer safe.  Once finished
we de-aco-ize the event and return into the original means.