Fix munmap in large http chains.
The http chain processor can sometimes under-use large blocks in a
bchain.  If they are large enough to qualify for mmap, they must
later be munmap'd instead of freed.  We used the `allocd` field
to track how large they were, but if we under-use the block, the
allocd will be forced down to match size (to ensure no one uses
the remainder) and then subsequent munmap will provide an
insufficient length to the call leaving some mmap'd memory stranded.
This fix uses a separate struct field (that is never modified) to
track the mmap size so that the corresponding munmap is guaranteed
to match.