Track the waitfor ping eventer explicitly and prevent double fire.
If two things are queued immediately for a coroutine that isn't in a
resumable state, it could register the ping twice which would cause
a lingering event to fire with out waiter.  By tracking the ping
eventer within the queue, we prevent a double ping.  Additionally,
we use this ping eventer to even more carefully remove it if there
was somehow a timeout fire race.