Make mtev_skiplist Safer If Misused (#488)
There are many functions in mtev_skiplist that will crash if passed a
null skiplist. Rather than crash, we should check for this and return a
failure.

Some cases, however are unrecoverable/we shouldn't try to recover from.
In these, we should actively assert instead of just returning or crashing.