Intern unswole (#493)
Several safety elements in slab allocation.

Two specific cases are handled wherein allocations during a
compaction could witness no available space, so we should compact
(which should be close to free if we were compacting prior) and then
attempt the freeslots escalation with the compaction lock.

The second issue is that we introduce the concept of views such that
when a node is plucked and moved into another slot we change the "view"
of the pool.  When we attempt out borrows and we fail, if we notice the
view has changed since we embarked, we should retry.