Fix null dereference caused by in mtev_memory_fini_thread.
We were using a boolean to determine if the asynch thread should
perform cleanup which could cause it to perform cleanup before the
last asynch reclamation job was enqueued.  Change the boolean to a
pointer representing what we know to be the last reclamation job
and trigger the termination on observation of that element.