Fix reading of chunked client request data.
Calculate minimal satisfied read correctly for request side chunked encoding.
We were not ensuring we had enough space for the trailing \r\n on the
chunked frame and large payloads that didn't otherwise fit in the
existing chain would be processed before we read the trailing \r\n and
then fail decoding. This change makes sure we size in two extra bytes
and ensures we've read them before we process the chunk.