Provide thread-safe access to http append methods.
This is an interesting one. It really makes almost no sense at all
to allow thread-safe access to an http output as it is unclear why
it would be useful to have undefined ordering of chunks of an HTTP
response, but I could see the case that if you were streaming JSON
(multi objects) and appended an object at a time atomically from
different threads it would work.

The is really to make sure we don't leak memory or crash if someone
tries to do this as the output bchain list manipulation could go
bad.