Make aco_resume work in ASAN.
aco_resume will step on the stack (as it intentionally overwrites the
red-zones). Makeing the aco_resume no_sanitize("address") isn't
sufficient as it calls memcpy() internally and that triggers the
stack-buffer-underflow condition.  This patch fixes some typos in
ASAN detection and then leverages that to replace memcpy with an
inline for loop which allows the no_sanitize("address") to work as
intended.