Implement flatbuffer (and json) logging.
Make logging internally all flatbuffer based. Provide a "format"
option for material loggers (those with ops).  Format defaults to
plain (which is unchanged behavior).  We also support "flatbuffer"
and "json" format.  Flatbuffer is not framed so it makes little
since to use it outside of a framed loggers. "jlog" and "memory"
types are framed, "file" is not.