Two-stage the predicate matching to avoid vprintf.
If the predicate for a filter doesn't include "message" then
we can do the evaluation without constructing the message and
it turns out that the vprintf can be quite expensive.  This will
break the predicate matching into two phases with the first phase
bailing early with an indicator if it cannot complete due to
needing the message constructed.

This bumps non-message predicates evaluation to 5.6mm/sec/core
on my test machine (from 2.6mm/sec/core).