integer arithmetic was correct and was returning the exact encode size (including padding only when necessary).
the original bug was likely with one of the hard coded encode/decode calculations elsewhere in the codebase and has been fixed by the other changes in this branch that are calling `mtev_b64_encode_len` and `mtev_b64_max_decode_len` directly.

`mtev_b64_max_decode_len` is returned to it's original state and can still return up the two extra bytes due to the potential for `=` padding in the encoded string.