Fix mtev_dyn_buffer_add_vprintf Bug (#919)
There was a bug in mtev_dyn_buffer_add_vprintf where we could call vsnprintf with a va_list twice without a copy. This has undefined behavior, as the call to vsnprintf is destructive. Fixed it so that it does a copy now. However, that copy could have performance impacts, so also added a new function, mtev_dyn_buffer_maybe_add_vprintf, that will not try to increase the buffer size and requires the caller to handle it in order to avoid the extra copy.