Add The Ability To Remove Kafka Connections (#952)
Add new functions to allow shutting down a specific consumer, a specific producer, or every active connection.
Added the ability to fetch a list of consumers or producers.
Added the ability to assign a unique UUID to each connection.
Fixed some bugs in the Kafka implementation that were exposed by the above work.

The code to close a connection is done after every polling cycle to reduce overhead - we build a list of consumers or produces to close, then remove them from the active list. The cleanup will then happen asynchronously and call a callback when the close is done. This is done to lower the overhead - removing connections is rare, and we don't want the work of doing so to interfere with the hot path of polling the consumers.